refactor: 조건별 게임 결과 확인을 위한 ResultRule 생성 및 책임 분리
