+++ src/main/java/blackjack/domain/card/Cards.java
@@ -5,13 +5,12 @@ import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
 
-import static blackjack.domain.card.CardDeck.NUMBER_OF_FIRST_CARDS;
-import static blackjack.domain.result.ResultType.BUST;
+import static blackjack.domain.result.ResultRule.BLACKJACK_SCORE;
 
 public class Cards {
     private static final int ACE_VALUE_DIFFERENCE = 10;
 
-    private List<Card> cards = new ArrayList<>();
+    private final List<Card> cards = new ArrayList<>();
 
     public void add(Card card) {
         this.cards.add(card);
@@ -26,7 +25,7 @@ public class Cards {
 
     private int handleAce(int sum) {
         for (Card card : cards) {
-            if (sum <= BUST) {
+            if (sum <= BLACKJACK_SCORE) {
                 break;
             }
 
@@ -45,14 +44,6 @@ public class Cards {
 
     public int size() {
         return cards.size();
-    }
-
-    public boolean isBlackJack() {
-        return computeScore() == BUST && cards.size() == NUMBER_OF_FIRST_CARDS;
-    }
-
-    public boolean isBust() {
-        return computeScore() > BUST;
     }
 
     @Override
